+++ src/components/subcircuits/chip.h
@@ -65,13 +65,13 @@ class Chip : public Component, public eElement
 
     protected:
         virtual Pin* addPin( QString id, QString type, QString label,
-                             int pos, int xpos, int ypos, int angle, int length=8, int space=0 ){return nullptr;}
+                             int pos, int xpos, int ypos, int angle, int length=8, double space=0 ){return nullptr;}
 
         virtual Pin* updatePin( QString id, QString type, QString label,
-                                int pos, int xpos, int ypos, int angle, int space=0  ){return nullptr;}
+                                int pos, int xpos, int ypos, int angle, double space=0  ){return nullptr;}
 
         virtual void addNewPin( QString id, QString type, QString label,
-                        int pos, int xpos, int ypos, int angle, int length=8, int space=0 );
+                        int pos, int xpos, int ypos, int angle, int length=8, double space=0 );
 
         void initPackage(  QString pkgStr  );
